requirements.txt
tweepy
python-dotenv
schedule
bottle
